[wmco] Move reusable Windows instances retrieval code
This commit modifies the method previously parseInstances to
reuse it in the CSR controller without having to create a
newConfigmapRenconciler object.
InstanceConfigMap constant and this method are moved
to a new package wiparser to avoid inverted dependency when
using them in the csr pkg.
Additionally a function getInstances() is added to return
the list of BYOH Windows instances without having to query
the configMap resource from the csr package.
